exclusão de arquivos de imagem

+++ cadastro-metas/excluir.php
@@ -5,14 +5,25 @@ include('../config/conexao_pdo.php');
 $codigo = $_GET['codigo'];
 $usuario = $_SESSION['usuario_codigo'];
 
-$sql = "DELETE FROM meta WHERE meta_codigo = :codigo and meta_usuario = :usuario";
+$sql = "SELECT meta_imagem FROM meta WHERE meta_codigo = :codigo and meta_usuario = :usuario";
 $stmt = $pdo->prepare($sql);
 $stmt->bindValue(':codigo', $codigo);
 $stmt->bindValue(':usuario', $usuario);
 $stmt->execute();
-
-
-echo "<br>Excluido com sucesso!";
-echo '<script>alertaSucesso("Cadastrado com sucesso!","lista.php")</script>';
-
+$nome_imagem = $stmt->fetch();
+$nome_imagem = 'imagens/' . $nome_imagem['meta_imagem'];
+if(file_exists($nome_imagem)){
+	if(unlink($nome_imagem)){
+		$sql = "DELETE FROM meta WHERE meta_codigo = :codigo and meta_usuario = :usuario";
+		$stmt = $pdo->prepare($sql);
+		$stmt->bindValue(':codigo', $codigo);
+		$stmt->bindValue(':usuario', $usuario);
+		$stmt->execute();
+		echo '<script>alertaSucesso("Excluído com sucesso!","lista.php")</script>';
+	}else{
+		echo '<script>alertaErro("Não foi possível excluir imagem da meta!",true)</script>';
+	}
+}else{
+	echo '<script>alertaErro("Imagem não encontrada!",true)</script>';
+}	
 ?>
